+++ miden-crypto/src/dsa/falcon512_poseidon2/keys/secret_key.rs
@@ -16,7 +16,9 @@ use super::{
 };
 use crate::{
     Word,
-    dsa::falcon512_rpo::{LOG_N, SK_LEN, hash_to_point::hash_to_point_rpo256, math::ntru_gen},
+    dsa::falcon512_poseidon2::{
+        LOG_N, SK_LEN, hash_to_point::hash_to_point_poseidon2, math::ntru_gen,
+    },
     hash::blake::Blake3_256,
     utils::zeroize::{Zeroize, ZeroizeOnDrop},
 };
@@ -149,7 +151,7 @@ impl SecretKey {
         let nonce = Nonce::deterministic();
 
         let h = self.compute_pub_key_poly();
-        let c = hash_to_point_rpo256(message, &nonce);
+        let c = hash_to_point_poseidon2(message, &nonce);
         let s2 = self.sign_helper(c, rng);
 
         Signature::new(nonce, h, s2)
@@ -169,7 +171,9 @@ impl SecretKey {
     /// These changes make the signature algorithm compliant with the reference implementation.
     #[cfg(test)]
     pub fn sign_with_rng_testing<R: Rng>(&self, message: &[u8], rng: &mut R) -> Signature {
-        use crate::dsa::falcon512_rpo::{hash_to_point::hash_to_point_shake256, tests::ChaCha};
+        use crate::dsa::falcon512_poseidon2::{
+            hash_to_point::hash_to_point_shake256, tests::ChaCha,
+        };
 
         let nonce = Nonce::random(rng);