+++ miden-crypto/src/dsa/mod.rs
@@ -2,4 +2,4 @@
 
 pub mod ecdsa_k256_keccak;
 pub mod eddsa_25519_sha512;
-pub mod falcon512_rpo;
+pub mod falcon512_poseidon2;