+++ miden-crypto/src/ies/mod.rs
@@ -47,8 +47,8 @@ use thiserror::Error;
 pub enum IesScheme {
     K256XChaCha20Poly1305 = 0,
     X25519XChaCha20Poly1305 = 1,
-    K256AeadRpo = 2,
-    X25519AeadRpo = 3,
+    K256AeadPoseidon2 = 2,
+    X25519AeadPoseidon2 = 3,
 }
 
 impl TryFrom<u8> for IesScheme {
@@ -57,8 +57,8 @@ impl TryFrom<u8> for IesScheme {
         match value {
             0 => Ok(IesScheme::K256XChaCha20Poly1305),
             1 => Ok(IesScheme::X25519XChaCha20Poly1305),
-            2 => Ok(IesScheme::K256AeadRpo),
-            3 => Ok(IesScheme::X25519AeadRpo),
+            2 => Ok(IesScheme::K256AeadPoseidon2),
+            3 => Ok(IesScheme::X25519AeadPoseidon2),
             _ => Err(IesError::UnsupportedScheme),
         }
     }
@@ -81,8 +81,8 @@ impl IesScheme {
         match self {
             IesScheme::K256XChaCha20Poly1305 => "K256+XChaCha20-Poly1305",
             IesScheme::X25519XChaCha20Poly1305 => "X25519+XChaCha20-Poly1305",
-            IesScheme::K256AeadRpo => "K256+AeadRpo",
-            IesScheme::X25519AeadRpo => "X25519+AeadRpo",
+            IesScheme::K256AeadPoseidon2 => "K256+AeadPoseidon2",
+            IesScheme::X25519AeadPoseidon2 => "X25519+AeadPoseidon2",
         }
     }
 }