+++ miden-crypto/src/lib.rs
@@ -142,10 +142,10 @@ pub trait SequentialCommit {
 
     /// Computes the commitment to the object.
     ///
-    /// The default implementation of this function uses RPO256 hash function to hash the sequence
-    /// of elements returned from [Self::to_elements()].
+    /// The default implementation of this function uses Poseidon2 hash function to hash the
+    /// sequence of elements returned from [Self::to_elements()].
     fn to_commitment(&self) -> Self::Commitment {
-        hash::rpo::Rpo256::hash_elements(&self.to_elements()).into()
+        hash::poseidon2::Poseidon2::hash_elements(&self.to_elements()).into()
     }
 
     /// Returns a representation of the object as a sequence of fields elements.